refactor: Reference 파일에서 익스텐션 분리
